sprint_2_langchain_rag/pfsp_config_copilot/requirements.txt
streamlit>=1.44.0,<2.0.0
langchain>=1.2.0,<2.0.0
langchain-openai>=1.0.1,<2.0.0
langchain-community>=0.4.1,<1.0.0
langchain-chroma>=0.2.0,<1.0.0
langchain-text-splitters>=0.3.0,<1.0.0
chromadb>=1.0.0,<2.0.0
openai>=2.0.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.12.0,<3.0.0
tiktoken>=0.9.0,<1.0.0
pytest>=8.0.0,<9.0.0
